Move play function into YouTube folder
